type-c-service: UCSI PPM implementation
