+++ examples/std/src/bin/type_c/external.rs
@@ -1,15 +1,55 @@
 //! Low-level example of external messaging with a simple type-C service
 use embassy_executor::{Executor, Spawner};
-use embedded_services::type_c::{ControllerId, external};
+use embassy_time::Timer;
+use embedded_services::{
+    GlobalRawMutex, power,
+    type_c::{ControllerId, external},
+};
 use embedded_usb_pd::GlobalPortId;
 use log::*;
 use static_cell::StaticCell;
+use std_examples::type_c::mock_controller;
+use type_c_service::wrapper::backing::BackingDefaultStorage;
+
+const CONTROLLER0: ControllerId = ControllerId(0);
+const PORT0: GlobalPortId = GlobalPortId(0);
+const POWER0: power::policy::DeviceId = power::policy::DeviceId(0);
+
+#[embassy_executor::task]
+async fn controller_task() {
+    static STATE: StaticCell<mock_controller::ControllerState> = StaticCell::new();
+    let state = STATE.init(mock_controller::ControllerState::new());
+
+    static BACKING_STORAGE: StaticCell<BackingDefaultStorage<1, GlobalRawMutex>> = StaticCell::new();
+    let backing_storage = BACKING_STORAGE.init(BackingDefaultStorage::new());
+    let backing = backing_storage.get_backing().expect("Failed to create backing storage");
+
+    static WRAPPER: StaticCell<mock_controller::Wrapper> = StaticCell::new();
+    let controller = mock_controller::Controller::new(state);
+    let wrapper = WRAPPER.init(mock_controller::Wrapper::new(
+        embedded_services::type_c::controller::Device::new(CONTROLLER0, &[PORT0]),
+        [power::policy::device::Device::new(POWER0)],
+        embedded_services::cfu::component::CfuDevice::new(0x00),
+        backing,
+        controller,
+        crate::mock_controller::Validator,
+    ));
+
+    wrapper.register().await.unwrap();
+    loop {
+        if let Err(e) = wrapper.process_next_event().await {
+            error!("Error processing wrapper: {e:#?}");
+        }
+    }
+}
 
 #[embassy_executor::task]
 async fn task(_spawner: Spawner) {
     info!("Starting main task");
     embedded_services::init().await;
 
+    // Allow the controller to initialize and register itself
+    Timer::after_secs(1).await;
     info!("Getting controller status");
     let controller_status = external::get_controller_status(ControllerId(0)).await.unwrap();
     info!("Controller status: {controller_status:?}");
@@ -30,6 +70,17 @@ async fn task(_spawner: Spawner) {
 
     info!("Setting retimer compliance");
     external::port_set_rt_compliance(GlobalPortId(0)).await.unwrap();
+
+    info!("Setting max sink voltage");
+    external::set_max_sink_voltage(GlobalPortId(0), Some(5000))
+        .await
+        .unwrap();
+
+    info!("Clearing dead battery flag");
+    external::clear_dead_battery_flag(GlobalPortId(0)).await.unwrap();
+
+    info!("Reconfiguring retimer");
+    external::reconfigure_retimer(GlobalPortId(0)).await.unwrap();
 }
 
 fn main() {
@@ -38,7 +89,8 @@ fn main() {
     static EXECUTOR: StaticCell<Executor> = StaticCell::new();
     let executor = EXECUTOR.init(Executor::new());
     executor.run(|spawner| {
-        spawner.must_spawn(type_c_service::task());
+        spawner.must_spawn(type_c_service::task(Default::default()));
         spawner.must_spawn(task(spawner));
+        spawner.must_spawn(controller_task());
     });
 }
