+++ type-c-service/src/service/controller.rs
@@ -1,0 +1,47 @@
+use embedded_services::{
+    debug, error,
+    type_c::{
+        external::{self, ControllerCommandData},
+        ControllerId,
+    },
+};
+
+use super::*;
+
+impl<'a> Service<'a> {
+    /// Process external controller status command
+    pub(super) async fn process_external_controller_status(
+        &self,
+        controller: ControllerId,
+    ) -> external::Response<'static> {
+        let status = self.context.get_controller_status(controller).await;
+        if let Err(e) = status {
+            error!("Error getting controller status: {:#?}", e);
+        }
+        external::Response::Controller(status.map(external::ControllerResponseData::ControllerStatus))
+    }
+
+    /// Process external controller sync state command
+    pub(super) async fn process_external_controller_sync_state(
+        &self,
+        controller: ControllerId,
+    ) -> external::Response<'static> {
+        let status = self.context.sync_controller_state(controller).await;
+        if let Err(e) = status {
+            error!("Error getting controller sync state: {:#?}", e);
+        }
+        external::Response::Controller(status.map(|_| external::ControllerResponseData::Complete))
+    }
+
+    /// Process external controller commands
+    pub(super) async fn process_external_controller_command(
+        &self,
+        command: &external::ControllerCommand,
+    ) -> external::Response<'static> {
+        debug!("Processing external controller command: {:#?}", command);
+        match command.data {
+            ControllerCommandData::ControllerStatus => self.process_external_controller_status(command.id).await,
+            ControllerCommandData::SyncState => self.process_external_controller_sync_state(command.id).await,
+        }
+    }
+}
